@@ -18,14 +18,10 @@ import static conf.Constants.SPEAKER_EMAIL;
 
 public class App {
 
-    /**
-     * NOTE: Future versions of Java will auto-box primitives.
-     *       Forced boxing will be removed.
-     */
-    //FIXME: 0. Replace redundant boxing
+    //NOTE: Future versions of Java will auto-box primitives.
+    //FIXME_1_: Redundant boxing of primitives - no longer needed
     public static final Integer ZERO_INTEGER = new Integer(0);
 
-    //FIXME: 11. Java 21 benefit
     /**
      * Orchestrates the conference creation, logistics and awards.
      * <p>
@@ -33,9 +29,10 @@ public class App {
      *
      * @param args Command line arguments (not used in this instance)
      */
+    //FIXME_11_: Java 21 benefit
     public static void main(String[] args) {
-        //FIXME: 0. Uncomment/Comment below code.
-        //showNullPointerException();
+        //FIXME_2_: Uncomment to see Helpful NullPointerExceptions - (comment out after).
+        showNullPointerException();
 
         Conference theConference = new Seeder().seed();
 
@@ -47,16 +44,18 @@ public class App {
         System.out.println("Indented Speaker email: ");
         System.out.println(SPEAKER_EMAIL);
 
-        //FIXME: 1. Use var instead
+        //FIXME_3_: Local Variable Type Inference: Use var instead
         Map<String, Integer> shirtCountMap = determineShirtCount(theConference);
         displayShirtCounts(shirtCountMap);
 
-        //FIXME: 1. Use var instead
+        //FIXME_3_: Local Variable Type Inference: Use var instead?
         Map<String, Integer> hatCountMap = null;
         hatCountMap = determineHatCount(theConference);
         displayHatCounts(hatCountMap);
 
         displayPaymentInvoicing(theConference);
+
+        calculateConferenceDiscount(theConference);
 
         displayBadgeCount(theConference);
 
@@ -86,8 +85,9 @@ public class App {
      * @param theConference - the current conference
      */
     static void displayBadgeCount(Conference theConference) {
-        //FIXME: 10. Convert to String Template FMT
-        //NOTE: FMT is a StringTemplate processor that allows String formatting.
+        //NOTE: String Templates are currently only available in Java 21 & 22
+        //      with enable-preview. String Templates will be revisited in a
+        //      future version of Java.
         StringBuffer message = new StringBuffer("");
         message.append("\n---------------------------------------------------------------");
         message.append("\n| " + StringUtils.rightPad("Participant type", 50) + " | " + StringUtils.leftPad("Count", 6) + " |");
@@ -115,8 +115,7 @@ public class App {
         for (Speaker speaker : theConference.getSpeakers()) {
             String shirtSize = speaker.getShirtSize();
             shirtCountMap.putIfAbsent(shirtSize, ZERO_INTEGER);
-            int currentCount = shirtCountMap.get(shirtSize);
-            shirtCountMap.put(shirtSize, currentCount + 1);
+            shirtCountMap.compute(shirtSize, (k, currentCount) -> currentCount + 1);
         }
         return shirtCountMap;
     }
@@ -134,8 +133,9 @@ public class App {
 
             int count = shirtCountMap.get(shirtSize);
 
-            //FIXME: 8. Convert to String Template STR
-            //NOTE: STR is the most basic StringTemplate processor.
+            //NOTE: String Templates are currently only available in Java 21 & 22
+            //      with enable-preview. String Templates will be revisited in a
+            //      future version of Java.
             String shirtSizes = "Shirt size: [" + shirtSize + "] " +
                     "-> Count: [" + count + "]";
 
@@ -154,8 +154,7 @@ public class App {
         for (Staff staff : theConference.getStaff()) {
             String hatSize = staff.getHatSize();
             hatCountMap.putIfAbsent(hatSize, ZERO_INTEGER);
-            int currentCount = hatCountMap.get(hatSize);
-            hatCountMap.put(hatSize, currentCount + 1);
+            hatCountMap.compute(hatSize, (k, currentCount) -> currentCount + 1);
         }
         return hatCountMap;
     }
@@ -170,8 +169,9 @@ public class App {
                 hatCountMap.values().stream().reduce(0, Integer::sum));
 
         for (String hatSize : hatCountMap.keySet()) {
-            //FIXME: 9. Convert to String Template RAW
-            //NOTE: RAW is a StringTemplate processor that can defer 'process'ing.
+            //NOTE: String Templates are currently only available in Java 21 & 22
+            //      with enable-preview. String Templates will be revisited in a
+            //      future version of Java.
             System.out.println("Hat size: [" + hatSize + "] " +
                     "-> Count: [" + hatCountMap.get(hatSize) + "]");
         }
@@ -187,7 +187,7 @@ public class App {
     static void displayPaymentInvoicing(Conference theConference) {
         double processingFee = 0.0D;
         for (Attendee attendee : theConference.getAttendees()) {
-            //FIXME: 2. Replace with switch expression
+            //FIXME_5_: Switch Expressions: Improve readability
             switch (attendee.getPaymentType()) {
                 case AMEX:
                     processingFee += 0.10D;
@@ -202,6 +202,25 @@ public class App {
             }
         }
         System.out.println("\nTotal payment-processing fees: USD[" + String.format("%,.2f", processingFee) + "]\n");
+    }
+
+    /**
+     * Calculate the discount that every attendee will receive.
+     *
+     * @param theConference
+     */
+    static void calculateConferenceDiscount(Conference theConference) {
+
+        double discount = Math.random();
+        System.out.println("Discount Alert!");
+        //FIXME_7_: Switch - Convert to primitive instance-of switch case with guards
+        if(discount <= 0.1d){
+            System.out.println("Nice! You get yourself a discount");
+        } else if (discount <= 0.2d){
+            System.out.println("Wow!! That is a nice discount");
+        } else {
+            System.out.println("Whoa!!! That is a whopper discount");
+        }
     }
 
     /**
@@ -259,7 +278,7 @@ public class App {
         }
 
         boolean allowedPersonFound = false;
-        //FIXME: 4. Replace the below with a switch pattern-matched instanceof
+        //FIXME_6_: Switch - Pattern-matched instanceof - improved readability
         for (AllowedPerson allowedPerson : allowedPersonWinner) {
             if (allowedPerson instanceof Attendee) {
                 Attendee attendee = (Attendee) allowedPerson;
@@ -297,14 +316,15 @@ public class App {
      * @param theConference - the current conference
      */
     static void displayRandomWinningSession(Conference theConference) {
-        //FIXME: 5b. Replace to a Record getter + upgrade to a toList() instead of Collectors.toList()
+        //FIXME_9_: Replace to a Record getter +
+        //       upgrade to a toList() instead of Collectors.toList()
         List<String> sessions = theConference.getSessions().stream().
                 map(Session::getSessionTitle).collect(Collectors.toList());
 
         String mostVotedSessionTitle = sessions.get(
                 ThreadLocalRandom.current().nextInt(0, sessions.size()));
 
-        //FIXME: 5c. Replace to a Record getter
+        //FIXME_9_: Replace to a Record getter
         Optional<Session> sessionObject = theConference.getSessions().stream()
                 .filter(session -> session.getSessionTitle().equals(mostVotedSessionTitle))
                 .findFirst();
@@ -324,19 +344,15 @@ public class App {
      */
     static void displaySessionDetails(Object object) {
 
-        //FIXME: 7. Use a record pattern
+        //FIXME_10_: Use a record pattern
         if (object instanceof Session) {
             Session session = (Session) object;
 
             String title = session.getSessionTitle();
             Speaker speaker = session.getMainSpeakerModerator();
 
-            String message = "\nThe most voted session: [" +
-                    title +
-                    "] by [" +
-                    speaker.firstName +
-                    " " +
-                    speaker.lastName + "]";
+            String message = String.format("\nThe most voted session: [%s] by [%s %s]",
+                    title, speaker.getFirstName(), speaker.getLastName()) ;
 
             System.out.println(message);
         }
@@ -352,13 +368,16 @@ public class App {
      *       "Helpful NullPointerException"
      */
     private static void showNullPointerException() {
-        Conference fakeConference = new Conference("fake", "fake", Year.now(), "fake");
+        Conference fakeConference = new Conference(
+                "fake", "fake", Year.now(), "fake");
         Session session = new Session(
-                "fake", "fake", new Speaker(null, "fake", "fake"));
+                "fake",
+                "fake",
+                new Speaker(null, "fake", "fake"));
         Set<Session> sessions = new HashSet<>();
         sessions.add(session);
         fakeConference.setSessions(sessions);
-        //FIXME: 5d. Replace to a Record getter
+        //FIXME_9_: Replace to a Record getter
         Object aSpeakerFirstNameLength =
                 ((Session) fakeConference.getSessions().
                         toArray()[0]).getMainSpeakerModerator().firstName.length();
