@@ -1,5 +1,7 @@
 package conf;
 
+import org.apache.commons.lang3.StringUtils;
+
 import java.time.Year;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -11,29 +13,21 @@ import java.util.concurrent.ThreadLocalRandom;
 
 import static conf.Constants.ATTENDEE_EMAIL;
 import static conf.Constants.SPEAKER_EMAIL;
-import static java.lang.StringTemplate.RAW;
-import static java.util.FormatProcessor.FMT;
 
 public class App {
-    //TODO: Add an example of a "switch statement with arrow syntax"
-    //TODO: Add an example of a "when clause" in a switch statement
-    //TODO: Add an example of "null pattern matching" in a switch block
 
-    /**
-     * NOTE: Future versions of Java will auto-box primitives.
-     *       Forced boxing will be removed.
-     */
-    //FIX ME: 0. Replace redundant boxing
+    //NOTE: Future versions of Java will auto-box primitives.
+    //FIXED 1: Replace redundant boxing
     public static final Integer ZERO_INTEGER = 0;
 
-    //FIX ME: 11. Java 21 benefit
     /**
      * Orchestrates the conference creation, logistics and awards.
      * <p>
      * NOTE: This method covers the usage of 'var' in Java
      */
+    //FIXED 11: Java 21 benefit
     void main() {
-        //FIX ME: 5. Uncomment/Comment below code.
+        //FIXED 2: Uncomment/Comment below code.
         //showNullPointerException();
 
         Conference theConference = new Seeder().seed();
@@ -46,16 +40,18 @@ public class App {
         System.out.println("Indented Speaker email: ");
         System.out.println(SPEAKER_EMAIL);
 
-        //FIX ME: 1. Use var instead
+        //FIXED 3: Use var instead
         var shirtCountMap = determineShirtCount(theConference);
         displayShirtCounts(shirtCountMap);
 
-        //FIX ME: 1. Use var instead
+        //FIXED 3: Use var instead
         var hatCountMap = determineHatCount(theConference);
 
         displayHatCounts(hatCountMap);
 
         displayPaymentInvoicing(theConference);
+
+        calculateConferenceDiscount(theConference);
 
         displayBadgeCount(theConference);
 
@@ -86,20 +82,33 @@ public class App {
      */
     static void displayBadgeCount(Conference theConference) {
 
-        //FIX ME: 10. Convert to String Template FMT
-        //NOTE: FMT is a StringTemplate processor that allows String formatting.
-        String message = FMT."""
-        ---------------------------------------------------------------
-        | Participant type                                   |  Count |
-        ===============================================================
-        | %-50s\{Speaker.class.getSimpleName()} | %6d\{theConference.getSpeakers().size()} |
-        | %-50s\{Attendee.class.getSimpleName()} | %6d\{theConference.getAttendees().size()} |
-        | %-50s\{Staff.class.getSimpleName()} | %6d\{theConference.getStaff().size()} |
-        | %-50s\{VendorSponsor.class.getSimpleName()} | %6d\{theConference.getVendorSponsors().size()} |
-        ===============================================================
-        | Total allowed people:                              | %6d\{determineBadgeCount(theConference)} |
-        ---------------------------------------------------------------
-        """;
+        //NOTE: String Templates are currently only available in Java 21 & 22
+        //      with enable-preview. String Templates will be revisited in a
+        //      future version of Java.
+        //JAVA22: FMT 'was' a StringTemplate processor that allowed String formatting.
+//        String message = FMT."""
+//        ---------------------------------------------------------------
+//        | Participant type                                   |  Count |
+//        ===============================================================
+//        | %-50s\{Speaker.class.getSimpleName()} | %6d\{theConference.getSpeakers().size()} |
+//        | %-50s\{Attendee.class.getSimpleName()} | %6d\{theConference.getAttendees().size()} |
+//        | %-50s\{Staff.class.getSimpleName()} | %6d\{theConference.getStaff().size()} |
+//        | %-50s\{VendorSponsor.class.getSimpleName()} | %6d\{theConference.getVendorSponsors().size()} |
+//        ===============================================================
+//        | Total allowed people:                              | %6d\{determineBadgeCount(theConference)} |
+//        ---------------------------------------------------------------
+//        """;
+        StringBuffer message = new StringBuffer("");
+        message.append("\n---------------------------------------------------------------");
+        message.append("\n| " + StringUtils.rightPad("Participant type", 50) + " | " + StringUtils.leftPad("Count", 6) + " |");
+        message.append("\n===============================================================");
+        message.append("\n| " + StringUtils.rightPad(Speaker.class.getSimpleName(), 50) + " | " + StringUtils.leftPad("" + theConference.getSpeakers().size(), 6) + " |");
+        message.append("\n| " + StringUtils.rightPad(Attendee.class.getSimpleName(), 50) + " | " + StringUtils.leftPad("" + theConference.getAttendees().size(), 6) + " |");
+        message.append("\n| " + StringUtils.rightPad(Staff.class.getSimpleName(), 50) + " | " + StringUtils.leftPad("" + theConference.getStaff().size(), 6) + " |");
+        message.append("\n| " + StringUtils.rightPad(VendorSponsor.class.getSimpleName(), 50) + " | " + StringUtils.leftPad("" + theConference.getVendorSponsors().size(), 6) + " |");
+        message.append("\n===============================================================");
+        message.append("\n| " + StringUtils.rightPad("Total allowed people: ", 50) + " | " + StringUtils.leftPad("" + determineBadgeCount(theConference), 6) + " |");
+        message.append("\n---------------------------------------------------------------");
 
         System.out.println(message);
     }
@@ -115,8 +124,7 @@ public class App {
         for (Speaker speaker : theConference.getSpeakers()) {
             String shirtSize = speaker.getShirtSize();
             shirtCountMap.putIfAbsent(shirtSize, ZERO_INTEGER);
-            int currentCount = shirtCountMap.get(shirtSize);
-            shirtCountMap.put(shirtSize, currentCount + 1);
+            shirtCountMap.compute(shirtSize, (k, currentCount) -> currentCount + 1);
         }
         return shirtCountMap;
     }
@@ -134,10 +142,14 @@ public class App {
 
             int count = shirtCountMap.get(shirtSize);
 
-            //FIX ME: 8. Convert to String Template STR
-            //NOTE: STR is the most basic StringTemplate processor.
-            String shirtSizes =
-                    STR. "Shirt size: [\{ shirtSize }] -> Count: [\{ count }]" ;
+            //NOTE: String Templates are currently only available in Java 21 & 22
+            //      with enable-preview. String Templates will be revisited in a
+            //      future version of Java.
+            String shirtSizes = "Shirt size: [" + shirtSize + "] " +
+                    "-> Count: [" + count + "]";
+            //JAVA22: STR 'was' the most basic StringTemplate processor.
+//            String shirtSizes =
+//                    STR. "Shirt size: [\{ shirtSize }] -> Count: [\{ count }]" ;
 
             System.out.println(shirtSizes);
         }
@@ -154,8 +166,7 @@ public class App {
         for (Staff staff : theConference.getStaff()) {
             String hatSize = staff.getHatSize();
             hatCountMap.putIfAbsent(hatSize, ZERO_INTEGER);
-            int currentCount = hatCountMap.get(hatSize);
-            hatCountMap.put(hatSize, currentCount + 1);
+            hatCountMap.compute(hatSize, (k, currentCount) -> currentCount + 1);
         }
         return hatCountMap;
     }
@@ -170,10 +181,14 @@ public class App {
                 hatCountMap.values().stream().reduce(0, Integer::sum));
 
         for (String hatSize : hatCountMap.keySet()) {
-            //FIX ME: 9. Convert to String Template RAW
-            //NOTE: RAW is a StringTemplate processor that can defer 'process'ing.
-            StringTemplate message = RAW."Hat size: [\{hatSize}] -> Count: [\{hatCountMap.get(hatSize)}]";
-            System.out.println(STR.process(message));
+            //NOTE: String Templates are currently only available in Java 21 & 22
+            //      with enable-preview. String Templates will be revisited in a
+            //      future version of Java.
+            //JAVA22: RAW 'was' a StringTemplate processor that can defer 'process'ing.
+//            StringTemplate message = RAW."Hat size: [\{hatSize}] -> Count: [\{hatCountMap.get(hatSize)}]";
+//            System.out.println(STR.process(message));
+            System.out.println("Hat size: [" + hatSize + "] " +
+                    "-> Count: [" + hatCountMap.get(hatSize) + "]");
         }
     }
 
@@ -187,7 +202,7 @@ public class App {
     static void displayPaymentInvoicing(Conference theConference) {
         double processingFee = 0.0D;
         for (Attendee attendee : theConference.getAttendees()) {
-            // FIX ME: 2. Replace with switch expression
+            // FIXED 5: Replace with switch expression
             processingFee += switch (attendee.getPaymentType()) {
                 case AMEX -> 0.10D;
                 case VISA, MASTERCARD -> 0.08D;
@@ -195,6 +210,25 @@ public class App {
             };
         }
         System.out.println("\nTotal payment-processing fees: USD[" + String.format("%,.2f", processingFee) + "]\n");
+    }
+
+    /**
+     * Calculate the discount that every attendee will receive.
+     *
+     * @param theConference
+     */
+    static void calculateConferenceDiscount(Conference theConference) {
+        double discount = Math.random();
+        System.out.println("Discount Alert!");
+        //FIXED 7: Switch - Convert to primitive instance-of switch case with guards
+        switch (discount) {
+            case double d when d <= 0.1d ->
+                System.out.println("Nice! You get yourself a discount");
+            case double d when d <= 0.2d ->
+                System.out.println("Wow!! That is a nice discount");
+            default ->
+                System.out.println("Whoa!!! That is a whopper discount");
+        }
     }
 
     /**
@@ -251,7 +285,7 @@ public class App {
             }
         }
 
-        //FIX ME: 4. Replace the below with a switch pattern-matched instanceof
+        //FIXED 6: Replace the below with a switch pattern-matched instanceof
         for (AllowedPerson allowedPerson : allowedPersonWinner) {
             switch (allowedPerson) {
                 case Attendee attendee -> System.out.println("Winner is an attendee: " +
@@ -280,7 +314,7 @@ public class App {
      * @param theConference - the current conference
      */
     static void displayRandomWinningSession(Conference theConference) {
-        //FIX ME: 5. Replace to a Record getter and
+        //FIXED 9: Replace to a Record getter and
         //          upgrade to a toList() instead of Collectors.toList()
         List<String> sessions = theConference.getSessions().stream().
                 map(Session::sessionTitle).toList();
@@ -288,7 +322,7 @@ public class App {
         String mostVotedSessionTitle = sessions.get(
                 ThreadLocalRandom.current().nextInt(0, sessions.size()));
 
-        //FIX ME: 5. Replace to a Record getter
+        //FIXED 9: Replace to a Record getter
         Optional<Session> sessionObject = theConference.getSessions().stream()
                 .filter(session -> session.sessionTitle().equals(mostVotedSessionTitle))
                 .findFirst();
@@ -306,14 +340,11 @@ public class App {
      * @param object An object that is intended to be a
      *               Session instance
      */
-    //TODO: Show a nested deconstruction of a record:
-    //  1. Convert AllowedPerson to be a sealed interface
-    //  2. Make Speaker a record type
-    //  3. Deconstruct Speaker and print "Congrats speaker.firstName!!!"
     static void displaySessionDetails(Object object) {
-        //FIX ME: 7. Use a record pattern
+        //FIXED 10: Use a record pattern
         if (object instanceof Session(String title, String _, Speaker speaker)) {
-            String message = STR. "\nThe most voted session: [\{ title }] by [\{ speaker.firstName } \{ speaker.lastName }]" ;
+            String message = String.format("\nThe most voted session: [%s] by [%s %s]",
+                    title, speaker.getFirstName(), speaker.getLastName()) ;
             System.out.println(message);
         }
     }
@@ -328,11 +359,15 @@ public class App {
      *       "Helpful NullPointerException"
      */
     private static void showNullPointerException() {
-        Conference fakeConference = new Conference("fake", "fake", Year.now(), "fake");
-        Session session = new Session("fake", "fake", new Speaker(null, "fake", "fake"));
+        Conference fakeConference = new Conference(
+                "fake", "fake", Year.now(), "fake");
+        Session session = new Session(
+                "fake",
+                "fake",
+                new Speaker(null, "fake", "fake"));
         Set<Session> sessions = Set.of(session);
         fakeConference.setSessions(sessions);
-        //FIX ME: 5. Replace to a Record getter
+        //FIXED 9: Replace to a Record getter
         Object aSpeakerFirstNameLength =
                 ((Session) fakeConference.getSessions().
                         toArray()[0]).mainSpeakerModerator().firstName.length();
